Transparency support experiment
This is badly unstable and awful :D
For some reason when a PNG is converted into a DIBv5
the transparency is lost and replaced with a black background,
and artifacts are added to the image (near edges of shapes).
This should work because I'm literally using the exact same
format as Mozilla when it copies an image to the clipboard.
But for me it doesn't work ! I have no idea why !
For some reason if in the picture only a single *one* pixel
has RGB values different from 0 with alpha not equal to 255,
every transparent pixels turn black. Because why not I guess.
So this is why I'm using this ugly hardcoded threshold for alpha
values. At least it works *approximately* with discord emotes...
wich is the main purpose of the software.