Fixed collision between previews loading threads.
Images' previews loading is now handled in a separate class
`PreviewsLoader`. The generation loop of previews is run
on another thread (`Task.Run`). The results' clearing
and building is run in an async which access is
controlled via a semaphore. This garantees that only
a signle thread can load previews at once.